+++ src/biz.dfch.CS.Abiquo.Client.Tests/v1/AbiquoClientTest.cs
@@ -40,7 +40,7 @@ namespace biz.dfch.CS.Abiquo.Client.Tests.v1
         private const string USERNAME = "ArbitraryUsername";
         private const string PASSWORD = "ArbitraryPassword";
         private const string ABIQUO_CLIENT_VERSION = "v1";
-        private const long TENANT_ID = 1;
+        private const int TENANT_ID = 1;
 
         [TestMethod]
         public void AbiquoClientVersionMatchesVersion3_8()
